pyproject.toml
[project]
name = "focal"
version = "0.1.0"
description = "Official implementation for Focal ICML 2025 paper."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.8.1",
    "bayesian-optimization==1.5.1",
    "diffusers==0.29.2",
    "equiadapt==0.1.2",
    "gsplat==1.4.0",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "omegaconf==2.3.0",
    "open-clip-torch>=2.29.0",
    "opencv-python>=4.12.0.88",
    "pandas==2.2.3",
    "pillow==11.0.0",
    "plyfile==1.1",
    "pycocotools==2.0.8",
    "pytorch-lightning==2.4.0",
    "scikit-learn==1.6.0",
    "scipy>=1.13.1",
    "segment-anything==1.0",
    "setuptools==72.1.0",
    "torch==2.4.1",
    "torchmetrics==0.11.4",
    "torchvision==0.19.1",
    "tqdm==4.66.5",
    "transformers>=4.48.1",
    "tyro==0.8.10",
    "xformers>=0.0.28.post1",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git", rev = "dca509fe793f601edb92606367a655c15ac00fdf" }
focal = { workspace = true }


[tool.setuptools.packages.find]
where = ["."]
include = ["focal", "focal*"]

[dependency-groups]
dev = [
    "focal",
]
